+++ lib/td_cx/auth/error_handler.ex
@@ -3,10 +3,8 @@ defmodule TdCx.Auth.ErrorHandler do
 
   import Plug.Conn
 
-  alias Jason, as: JSON
-
   def auth_error(conn, {type, _reason}, _opts) do
-    body = JSON.encode!(%{message: to_string(type)})
+    body = Jason.encode!(%{message: to_string(type)})
     send_resp(conn, 401, body)
   end
 end